临时解决编译报错的问题

+++ Source/Engine/Resource/resource_object.cpp
@@ -15,7 +15,7 @@ namespace volucris
     {
         if (asset.type != m_asset.type)
         {
-            V_LOG_ERROR(Engine, "try set invalid asset info to asset object. type {}, object type {}", asset.type, m_asset.type);
+            V_LOG_ERROR(Engine, "try set invalid asset info to asset object. type {}, object type {}", (int)asset.type, (int)m_asset.type);
         }
         else
         {
